Add flags argument to create_{buffer,surface}
This adds an additional argument to the create_{buffer,surface}. The
upper 16 bits of this number are global flags specified in wld.h, and
the lower 16 bits are implementation-specific flags.

Some drivers, like nouveau, need extra information about intended use
for a buffer in order to create it with the correct settings.